chore: Simplify build script to only use vite (and migrate to `@nextcloud/vite-config`)
