+++ package.json
@@ -2,16 +2,25 @@
   "name": "@nextcloud/password-confirmation",
   "version": "4.0.4",
   "description": "Password confirmation for Nextcloud",
-  "main": "./dist/main.js",
-  "browser": "./dist/main.js",
-  "types": "./dist/main.d.ts",
+  "main": "./dist/index.cjs",
+  "types": "./dist/index.d.ts",
+  "exports": {
+    ".": {
+      "types": "./dist/style.css",
+      "import": "./dist/index.mjs",
+      "require": "./dist/index.cjs"
+    },
+    "./style.css": "./dist/style.css",
+    "./dist/style.css": "./dist/style.css"
+  },
   "files": [
-    "dist"
+    "dist",
+    "CHANGELOG.md"
   ],
   "scripts": {
     "dev": "vite build --mode development",
     "watch": "vite build --mode development --watch",
-    "build": "vue-tsc && vite build && tsup src/main.ts --silent --dts-only",
+    "build": "vite build --mode production && mv dist/main.css dist/style.css",
     "l10n:extract": "node build/extract-l10n.mjs",
     "lint": "eslint --ignore-pattern dist/**/* .",
     "lint:fix": "eslint --fix --ignore-pattern dist/**/* ."
@@ -28,30 +37,27 @@
     "@nextcloud/l10n": "^1.6.0",
     "@nextcloud/router": "^2.0.0",
     "@nextcloud/vue": "^7.6.0",
-    "vue": "^2.7.10"
+    "vue": "^2.7.14"
   },
   "devDependencies": {
     "@nextcloud/browserslist-config": "^3.0.0",
-    "@nextcloud/eslint-config": "^8.2.1",
+    "@nextcloud/eslint-config": "^8.3.0",
+    "@nextcloud/vite-config": "^1.0.0-beta.19",
     "@types/gettext-parser": "^4.0.1",
     "@types/node": "^20.6.3",
     "@types/node-gettext": "^3.0.3",
-    "@vitejs/plugin-vue2": "^2.0.0",
-    "browserslist-to-esbuild": "^1.2.0",
     "gettext-extractor": "^3.5.4",
     "gettext-parser": "^7.0.0",
     "sass": "^1.54.8",
-    "tsup": "^7.2.0",
-    "typescript": "^4.8.2",
-    "vite": "^4.3.8",
-    "vue-tsc": "^1.0.0"
+    "typescript": "^5.0.2",
+    "vite": "^4.4.9"
   },
   "browserslist": [
     "extends @nextcloud/browserslist-config"
   ],
   "eslintConfig": {
     "extends": [
-      "@nextcloud"
+      "@nextcloud/eslint-config/typescript"
     ]
   },
   "engines": {
