+++ tsconfig.node.json
@@ -5,5 +5,5 @@
     "moduleResolution": "node",
     "allowSyntheticDefaultImports": true,
   },
-  "include": ["vite.config.ts"],
+  "include": ["vite.config.mts"],
 }
